test: add a successful direct command over SSH
The test looks for the environment variable `ketrew_test_ssh`; if not
defined it will use `"localhost"`, for the test to succeed it should be
a passwordless SSH host (the only command run on it is for now `ls /`).
Run the test like:

    ketrew_test_ssh=<Host> _obuild/ketrew-test/ketrew-test.asm [-no-color]

where `<Host>` is often an entry in your `.ssh/config` file.