test: add `cli.ml` (website workflow)
The new test ketrew-cli-test has a first (non-EDSL) workflow that can
compile the documentation, switch to the gh-pages branch, commit the new
version of the documentation, and got back to the original branch.
It can also fail in many ways (e.g. Git checkout forbidden because of
dirty tree) which show ketrew's handling of errors in dependencies.