lib: daemonize server before any Lwt threading
This is related to the issue
[`ocsigen/lwt#82`](https://github.com/ocsigen/lwt/issues/82):
One can call `Lwt_daemon.daemonize` only before the Lwt engine starts.

-  The function `Configuration.get_configuration_non_deferred_exn`
does the same thing as `Configuration.get_configuration` but without Lwt
(and may throw exceptions).
- The daemonization is moved to the module `Ketrew_command_line`.

Before this change the calls to `Lwt_preemptive.detach` (used to access
the database and for the `Dynlink` module) were hanging.