lib: fix and improve the Target graph clean-up
It used to be based only on dependencies; now it takes everything into
account. The clean-up algorithm is not “complete” but conservative: a
target that is not activable could resist the garbage colletion because
a target in its transitive sub-graph is still activated or running
(which should not be indefinite).