lib: implement the PBS backend + integration test
There is a new module `Ketrew_pbs` which implements LONG_RUNNING with
PBS/Torque. This was developed in conjunction with integration tests:
the test uses Ketrew to build a vagrant VM with Torque installed; then
the PBS setup is finalized, and then 2 PBS jobs are run (one of them
always fails, it is supposed to).