Rewrite the target automaton
This is a huge commit because it starts by revamping the internal representation
of targets, and rewrites the `State.t` types into a much more precise automaton
encoded in the type system.

The `Ketrew_engine` part has been mostly adapted but there are some `assert
false` here and there still. Most “internal” functionality is almost there,
following commits will keep going on cleaning up, documenting, and adapting the
UI.

The EDSL itself is moslty untouched.

The automatic tests are not very advanced but the the framework for testing
targets is on a decent shape.