Implement “staging area” for targets
When `Engine.add_targets` is called the targets are now stored in another
collection of the database, and then they are actually checked against other
targets and added to the main collection at the end of the following `step`.

The “killing-related” code was almost move to a sub-module to match the one for
adding.

The `basic-test` was updated, to matche the new slighly different behavior.

The protocol was simplified since the the server cannot reply which targets were
converted to pointers any more.