Switch build-system to OMake
The build of the documentation “works” but the documentation is broken in many
places because of the packing of the libraries.

The `OMakeroot` and `tools/OMakeLibrary.om` replace `Makefile`,
`tools/please.ml`, and `tools/_oasis.in`.
And  `tools/OMakeLibrary.om` is meant to be reusable and synchronized
with [`smondet/omake-hacks`](https://github.com/smondet/omake-hacks)
for use in other projects.

This also deletes the `.merlin`, just use `omake .merlin` to regenerate it.