Keep fetching passive targets
This moslty removes previous “debug” code that was stopping the update when
nothing seems to be happening (i.e. no target can activate a passive one); but
even in that case the passive targets can be killed (from the GUI itself or not)
so we need to keep updating the passive targets.