Create a typed “ring-log”
It is for now used in `Server`, `Engine`, and `Persistent_data`.

The logs, while in memory, can be dumped to files with the server's command-pipe
in JSON format or pretty-printed.

    echo "get-log:$PWD/some-file.log:txt" > _test_env/test-command.pipe

or

    echo "get-log:$PWD/some-file.json:json" > _test_env/test-command.pipe

The configuration parameter `log_path` of the server changes its semantics, it
is now a directory that contains periodic JSON dumps of the in-memory logs.

The debug printing (while daemonized) now goes to a `debug.txt` file in that
directory.