Make list-of-ids queries “deferrable”
Now there are new kinds of protocol messages: when results are too big the
server can “save” the result and return an ID of the result that can be queried
later.

The client for now just gets the first 500 results for the table.

The protocol is mostly backwards compatible since a new client can talk to an
older server.

This commit has the side effect of improving the logs that the server outputs at
every loop of the engine.