Implement basic protocol + UI for `Process_holder`
One can do basic things, like start an SSH connection, kill it, talk to the
`ASKPASS` program to send their password, and the status is displayed (based on
the status of the process group for now).

The UI is pretty ugly and the list of connection lacks reactivity (“Update list”
button always needed).