Group the Target processing Server logic into module
The intent of this patch is to make the code a bit terser and better
organized. To that end, some of the redundant prefixes ("answer_")
and suffixes ("_target") have been removed in favor of a Targets
module.

The label on 'id' Engine.kill has been removed so that this method
can be used in a curried style.