Introduce named-hosts in `Host` and `Host_io`
Now `host_io.t` is a map from names to `Host.t` that needs to be present
for all operations on `Host.t`s. Hence, most of the changes in this
patch are related to passing the argument `~host_io` around all the
code. This will be anyways useful for future work on scheduling/batching
of SSH connections.

This breaks backwards compatibility of the `LONG_RUNNING` module type.