Improve compilation of the target-table filters
Not `(not ...)` expression are explored further but it is still not
perfect because the protocol does not allow to express all the filters.

Changing the protocol to be complete is left for future work.

This fixes in particular the issue #288, the filter `(not (all))` gets
transformed to `` `Created_after max_float`` which is about 5.7 × 10^300
years in the future.