Add “new workflow” notifications to the WebUI
This adds a notification section/area to the UI.

For now only one kind of real notifications is implemented and displayed:
the acceptance of a new workflow to be added to the engine.
This visually fixes #247 since workflows show up instantly in the
notification area (they don't have to wait to be properly started to the
engine).

In this first version, the notification area is an `h3` section under
the client “tabs” this may change later.