Allow on-demand “debug-display” of structured-logs
One can now ask Ketrew to display on `stderr` a subset of the markup-logs:

    debug_log_functions=block_if_empty_at_most,create kdserver start

is for example going to print all the logs that have the “`function`” field
equal to `block_if_empty_at_most` or `create`.

It is also possible to use the `debug_log_modules` environment variable.
The final display condition is
`` `Ignore_case (`Or (modules @ functions))``.