Implement temporary “freezing” of node states
The idea is that when a node is in `` `Still_running`` or
`` `Tried_to_eval_condition` states the engine will check again their
state only after a given amount of time (default 30 seconds).

This makes the engine more “reactive” to new workflows: a slow SSH
connection on an existing long-running job won't slow down the bootstrap
of a new workflow coming in.